Add SMTPConnection test
Added a test if it's possible to open a connection to MX servers of a
domain.

Passing results of previous tests along -- we have just three tests so
no problem in hardcoding them and we don't have to do eg. the MX
resolving test again when trying to connect to the SMTP servers.